Change options, allow either emerge or cmake build
This is a pretty big change - it breaks the command line sort of.
Now you call out what you want it to do.  Usage is:
options:
    -3 Also build third party stuff
    -a Announce stuff as you go
    -c Build kstars via cmake (ONLY one of -c or -e can be used)
    -d Dry run only (just show what you are going to do)
    -e Build kstars via emerge
    -i Build libindi
    -s Skip brew (only use this if you know you already have them)

So, to build a complete emerge you would do:
    $0 -3aei

So, to build a complete cmake build you would do:
    $0 -3aci

If you add a -s it skips the brew steps - saves a few seconds